+++ packages/protocol/test/unit/protocol/control_layer/ProxyConnectionRequest.test.ts
@@ -1,7 +1,8 @@
 import assert from 'assert'
-
 import ValidationError from '../../../../src/errors/ValidationError'
-import { ProxyConnectionRequest, ControlMessage, toStreamID } from '../../../../src/exports'
+import ControlMessage from '../../../../src/protocol/control_layer/ControlMessage'
+import ProxyConnectionRequest from '../../../../src/protocol/control_layer/proxy_connection_request/ProxyConnectionRequest'
+import { toStreamID } from '../../../../src/utils/StreamID'
 import { ProxyDirection } from '../../../../src/utils/types'
 
 describe('ProxyConnectionRequest', () => {
