+++ packages/protocol/test/unit/protocol/control_layer/ProxyConnectionResponse.test.ts
@@ -1,7 +1,8 @@
 import assert from 'assert'
-
 import ValidationError from '../../../../src/errors/ValidationError'
-import { ProxyConnectionResponse, ControlMessage, toStreamID } from '../../../../src/exports'
+import ControlMessage from '../../../../src/protocol/control_layer/ControlMessage'
+import ProxyConnectionResponse from '../../../../src/protocol/control_layer/proxy_connection_response/ProxyConnectionResponse'
+import { toStreamID } from '../../../../src/utils/StreamID'
 import { ProxyDirection } from '../../../../src/utils/types'
 
 describe('ProxyPublishStreamConnectionResponse', () => {
