+++ packages/protocol/test/unit/protocol/message_layer/StreamMessage.test.ts
@@ -1,19 +1,15 @@
-import assert from 'assert'
-
-import sinon from 'sinon'
-
-import {
-    MessageRef,
-    MessageID,
-    EncryptedGroupKey,
-    toStreamID,
-    StreamPartIDUtils
-} from '../../../../src/exports'
-import ValidationError from '../../../../src/errors/ValidationError'
-import UnsupportedVersionError from '../../../../src/errors/UnsupportedVersionError'
-import { Serializer } from '../../../../src/Serializer'
-import StreamMessage, { ContentType, EncryptionType, StreamMessageType } from '../../../../src/protocol/message_layer/StreamMessage'
 import { toEthereumAddress } from '@streamr/utils'
+import assert from 'assert'
+import sinon from 'sinon'
+import UnsupportedVersionError from '../../../../src/errors/UnsupportedVersionError'
+import ValidationError from '../../../../src/errors/ValidationError'
+import EncryptedGroupKey from '../../../../src/protocol/message_layer/EncryptedGroupKey'
+import MessageID from '../../../../src/protocol/message_layer/MessageID'
+import MessageRef from '../../../../src/protocol/message_layer/MessageRef'
+import StreamMessage, { ContentType, EncryptionType, StreamMessageType } from '../../../../src/protocol/message_layer/StreamMessage'
+import { Serializer } from '../../../../src/Serializer'
+import { toStreamID } from '../../../../src/utils/StreamID'
+import { StreamPartIDUtils } from '../../../../src/utils/StreamPartID'
 
 const content = {
     hello: 'world',
