+++ packages/protocol/test/unit/protocol/tracker_layer/RelayMessageSerializerV2.test.ts
@@ -1,6 +1,5 @@
 import assert from 'assert'
-
-import { RelayMessage, RelayMessageSubType } from '../../../../src/exports'
+import RelayMessage, { RelayMessageSubType } from '../../../../src/protocol/tracker_layer/relay_message/RelayMessage'
 import TrackerMessage from '../../../../src/protocol/tracker_layer/TrackerMessage'
 
 const VERSION = 2
