+++ packages/protocol/test/unit/protocol/tracker_layer/StatusAckMessageSerializerV2.test.ts
@@ -1,7 +1,7 @@
 import assert from 'assert'
-
-import { StatusAckMessage, toStreamID } from '../../../../src/exports'
+import StatusAckMessage from '../../../../src/protocol/tracker_layer/status_ack_message/StatusAckMessage'
 import TrackerMessage from '../../../../src/protocol/tracker_layer/TrackerMessage'
+import { toStreamID } from '../../../../src/utils/StreamID'
 
 const VERSION = 2
